api/requirements.txt
aiofiles==0.4.0
aredis==1.1.3
bcrypt==3.1.6
cffi==1.11.5
entrypoints==0.3
flake8==3.7.1
httptools==0.0.11
manage.py==0.2.10
mccabe==0.6.1
multidict==4.5.2
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.0
PyJWT==1.7.1
python-dotenv==0.10.1
sanic==18.12.0
sanic-jwt==1.2.1
six==1.12.0
SQLAlchemy==1.2.17
ujson==1.35
uvloop==0.12.0
websockets==6.0
